pyproject.toml
[project]
name = "mcp-intro"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.12.0",
    "httpx>=0.28.1",
    "langchain-mcp-adapters>=0.1.13",
    "langchain-openai>=1.0.3",
    "langgraph>=1.0.3",
    "mcp[cli]>=1.8.0",
    "playwright>=1.40.0",
    "python-dotenv>=1.2.1",
]
